# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Omersia Storefront API",
        "description": "Documentation de l'API publique Storefront pour l'e-commerce Omersia",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Environnement local / base"
        }
    ],
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Lister toutes les catégories actives",
                "description": "Retourne toutes les catégories actives (enfants de la catégorie 'accueil') avec leurs traductions et le nombre de produits",
                "operationId": "getCategories",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Langue des traductions",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "fr",
                            "enum": [
                                "fr",
                                "en"
                            ]
                        }
                    },
                    {
                        "name": "parent_only",
                        "in": "query",
                        "description": "Retourner uniquement les catégories parentes",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des catégories avec comptage des produits",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "API key invalide ou manquante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiKeyError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api.key": []
                    }
                ]
            }
        },
        "/api/v1/categories/{slug}": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Détail d'une catégorie avec ses produits et sous-catégories",
                "description": "Retourne une catégorie par slug avec ses sous-catégories (jusqu'à 3 niveaux), sa catégorie parente et les produits paginés (20/page)",
                "operationId": "getCategoryBySlug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug de la catégorie dans la langue spécifiée",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "chemises-homme"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Langue des traductions",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "fr",
                            "enum": [
                                "fr",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie trouvée avec produits et sous-catégories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryDetailResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catégorie non trouvée pour ce slug et cette langue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "API key invalide ou manquante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiKeyError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api.key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CategoryDetailResponse": {
                "description": "Category payload with related products",
                "properties": {
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    }
                },
                "type": "object"
            },
            "CategoryListResponse": {
                "description": "List of categories",
                "properties": {
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object"
            },
...
            "NotFoundError": {
                "description": "Resource not found error (HTTP 404)",
                "properties": {
                    "message": {
                        "description": "Not found message",
                        "type": "string",
                        "example": "Resource not found"
                    }
                },
                "type": "object"
            },
            "ApiKeyError": {
                "description": "API key validation error",
                "properties": {
                    "message": {
                        "description": "API key error message",
                        "type": "string",
                        "example": "Invalid API key"
                    },
                    "error": {
                        "description": "Detailed error (e.g., expired, missing)",
                        "type": "string",
                        "example": "API key expired",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ProductTranslation": {
                "properties": {
                    "locale": {
                        "type": "string",
                        "example": "fr"
                    },
                    "name": {
                        "type": "string",
                        "example": "T-shirt Homme"
                    },
                    "slug": {
                        "type": "string",
                        "example": "t-shirt-homme"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "short_description": {
                        "type": "string",
                        "nullable": true
                    },
                    "meta_title": {
                        "type": "string",
                        "nullable": true
                    },
                    "meta_description": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CategoryTranslation": {
                "properties": {
                    "locale": {
                        "type": "string",
                        "example": "fr"
                    },
                    "name": {
                        "type": "string",
                        "example": "Vêtements"
                    },
                    "slug": {
                        "type": "string",
                        "example": "vetements"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "meta_title": {
                        "type": "string",
                        "nullable": true
                    },
                    "meta_description": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Vêtements",
                        "nullable": true
                    },
                    "slug": {
                        "type": "string",
                        "example": "vetements",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "example": "https://cdn.example.com/category.jpg",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://cdn.example.com/category.jpg",
                        "nullable": true
                    },
                    "count": {
                        "type": "integer",
                        "example": 12,
                        "nullable": true
                    },
                    "parent_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "position": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "translations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryTranslation"
                        }
                    },
                    "parent": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Category"
                            }
                        ],
                        "nullable": true
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object"
            },
            "ProductImage": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 34
                    },
                    "url": {
                        "type": "string",
                        "example": "https://cdn.example.com/products/image1.jpg"
                    },
                    "path": {
                        "type": "string",
                        "example": "products/image1.jpg",
                        "nullable": true
                    },
                    "alt": {
                        "type": "string",
                        "example": "Vue principale",
                        "nullable": true
                    },
                    "position": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_main": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "ProductVariantValue": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "value": {
                        "type": "string",
                        "example": "M"
                    },
                    "option": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Taille"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "OptionValue": {
                "properties": {
                    "option": {
                        "type": "string",
                        "example": "Taille"
                    },
                    "value": {
                        "type": "string",
                        "example": "M"
                    }
                },
                "type": "object"
            },
            "ProductVariant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 456
                    },
                    "sku": {
                        "type": "string",
                        "example": "TSHIRT-BLACK-M",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Noir / M",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "manage_stock": {
                        "type": "boolean",
                        "example": true
                    },
                    "stock_qty": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "compare_at_price": {
                        "type": "number",
                        "format": "float",
                        "example": 39.99,
                        "nullable": true
                    },
                    "on_sale": {
                        "type": "boolean",
                        "example": false
                    },
                    "values": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductVariantValue"
                        }
                    },
                    "option_values": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OptionValue"
                        }
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 123
                    },
                    "sku": {
                        "type": "string",
                        "example": "TSHIRT-BLACK-L",
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "example": "simple"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "manage_stock": {
                        "type": "boolean",
                        "example": true
                    },
                    "stock_qty": {
                        "type": "integer",
                        "example": 42,
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "compare_at_price": {
                        "type": "number",
                        "format": "float",
                        "example": 39.99,
                        "nullable": true
                    },
                    "on_sale": {
                        "type": "boolean",
                        "example": false
                    },
                    "has_variants": {
                        "type": "boolean",
                        "example": false
                    },
                    "from_price": {
                        "type": "number",
                        "format": "float",
                        "example": 19.99,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "T-shirt Homme",
                        "nullable": true
                    },
                    "slug": {
                        "type": "string",
                        "example": "t-shirt-homme",
                        "nullable": true
                    },
                    "short_description": {
                        "type": "string",
                        "nullable": true
                    },
                    "main_image_url": {
                        "type": "string",
                        "example": "https://cdn.example.com/products/image1.jpg",
                        "nullable": true
                    },
                    "translations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductTranslation"
                        }
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductImage"
                        }
                    },
                    "variants": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductVariant"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api.key": {
                "type": "apiKey",
                "description": "Frontend API Key (required for all requests)",
                "name": "X-API-KEY",
                "in": "header"
            },
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
...
    "security": [
        {
            "api.key": []
        }
    ]
}